implement automatic retry
